+++ src/components/DaySelection/index.tsx
@@ -144,7 +144,8 @@ export default function DaySelection({
                         {course.type === ScheduleBlockEventType.CustomEvent &&
                           course.where && (
                             <span className="course-row">
-                              {/* avoid showing full address is there a better way to format this? */}
+                              {/* avoid showing full address
+                                  is there a better way to format this? */}
                               {course.where.split(',')[0]}
                             </span>
                           )}
